Schemas separation
Se realiza la separación de los schemas en archivos separados, centralizandolos en la carpeta schemas,  para mejorar la mantenibilidad del código
