+++ qward/metrics/qiskit_metrics.py
@@ -12,7 +12,7 @@ from qward.metrics.types import MetricsType, MetricsId
 
 # Import schemas for structured data validation
 try:
-    from qward.metrics.schemas import (
+    from qward.schemas.qiskit_metrics_schema import (
         QiskitMetricsSchema,
         BasicMetricsSchema,
         InstructionMetricsSchema,
