+++ qward/metrics/types.py
@@ -15,6 +15,7 @@ class MetricsId(Enum):
     CIRCUIT_PERFORMANCE = "CIRCUIT_PERFORMANCE"
     LOC = "LOC"
     QUANTUM_HALSTEAD = "QUANTUM_HALSTEAD"
+    QC_UNDERSTANDABILITY = "QC_UNDERSTANDABILITY"
 
 
 class MetricsType(Enum):
